from 42

+++ charenge_00/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: oitsuki <oitsuki@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/28 05:47:09 by oitsuki           #+#    #+#             */
-/*   Updated: 2025/04/28 06:04:20 by oitsuki          ###   ########.fr       */
+/*   Updated: 2025/04/30 12:45:42 by oitsuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_strrchr(const char *s, int c)
 	{
 		if (s[len] == (char)c)
 			return ((char *)&s[len]);
-		if (len == c)
+		if (len == (size_t)c)
 			break ;
 		len--;
 	}
